--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,4 @@
 *.iml
 .tags
 .tags*
-node_modules
+node_modules/.bin