--- a/spec/resources/small/dom-node.html
+++ b/spec/resources/small/dom-node.html
@@ -1,3 +1,3 @@
-<div class="red"></div>
+<div class="red" onclick="stuff('<encode me or i'll break your image loading>')"></div>
 <div class="green"></div>
 <div class="blue"></div>