… README.md
… spec/dom-to-image.spec.js