I'm using this library in sort of a batch situation, where the user is sitting there waiting for it to finish, so I wanted to see if I could speed it up in some way. I tried two approaches - 
1) Since the end goal is having the DOM serialized to XHTML, then instead of actually cloning the elements, which is expensive, rather scan them and store the necessary information in a native object structure and serialize that.
2) Building on 1, use web workers to do the work in the middle of the pipeline (after the elements are scanned, and before the final Image is produced). Web workers are not allowed to touch the document or elements, so the goal was to get them out of the picture as early as possible and switch to the other thread.

I [benchmarked](https://codepen.io/anon/pen/MmORPO?editors=0011) each of these approaches along with the master branch against a fairly complicated DOM, in blocking and non-blocking flavors, where

Blocking is equivalent to
```
domtoimage.toPng(element).then(function() {
    domtoimage.toPng(element).then(function() {
        ...
    });
});
```

Non-blocking is equivalent to
```
Promise.all([
    domtoimage.toPng(element),
    domtoimage.toPng(element),
    ...
]).then(...);
```

Here are typical results I get from it on Chrome. "Original" refers to your master branch, and "Rewrite" and "RewriteWorkers" correspond to 1) and 2) above, respectively:
```
"Running tests..."
"blocking(singleThreaded(libs.domtoimageOriginal))"
"2.95 ops/sec"
"nonBlocking(singleThreaded(libs.domtoimageOriginal))"
"4.13 ops/sec"
"blocking(singleThreaded(libs.domtoimageRewrite))"
"4.71 ops/sec"
"nonBlocking(singleThreaded(libs.domtoimageRewrite))"
"9.16 ops/sec"
"blocking(multiThreaded(libs.domtoimageRewriteWorkers))"
"4.50 ops/sec"
"nonBlocking(multiThreaded(libs.domtoimageRewriteWorkers))"
"9.89 ops/sec"
```

Approach 1 gives a healthy improvement, 60% for blocking and 120% for non-blocking, but adding multithreading does not improve it any further. So it appears that actually cloning elements is relatively slow, while pulling the data off them and manipulating a native javascript object is quicker. Looking at CPU profiles of each one, it looks like multithreading doesn't help significantly because most of the work done in the middle of the pipeline is already non-blocking (XHR), so a single thread is able to perform nearly as well as a pool. The single thread is slightly faster in the blocking test since you're waiting for each iteration to finish, so there's no parallelism, only threading overhead. And it's slightly slower in the non-blocking test because the final serialization is a bit of blocking work that gives the threadpool a slight edge. However it's insignificant enough that this PR is for approach 1.