@tclausing 2017-05-11
I added a new approach that builds on 1) and works well. It works better for a batch of different DOMs by offering a separate "scan" function that just collects the DOM information and returns it. The consumer can then pass that object into one of the toXYZ functions for final rendering, but they can also immediately move on to the next DOM and start scanning it; they don't have to wait for rendering of the first DOM to finish. So it still benefits nicely from non-blocking effects during the remainder of the pipeline, and it actually gets a speedup nearly as big as the full non-blocking test above, but it's more helpful in real world situations. In my real world situation I would be going from
```
"blocking(singleThreaded(libs.domtoimageOriginal))"
"2.95 ops/sec"
```
to
```
"chunk(libs.domtoimageRewriteChunk)"
"8.32 ops/sec"
```
for a nice 182% improvement.