@sarath-sasikumar 2017-05-22
@tclausing , by how much is the approach 1 mentioned in the comment above improving the conversion performance.