@tclausing 2017-05-23
@sarath-sasikumar for a single conversion (or for a sequence of blocking conversion calls), the improvement was 60%. You can run the benchmark yourself (on chrome) [here](https://codepen.io/anon/pen/MmORPO?editors=0011), comparing "blocking(singleThreaded(libs.domtoimageOriginal))" to 
"blocking(singleThreaded(libs.domtoimageRewrite))".