@tsayen 2017-06-10
Hi. Sorry, I haven't had time to review those changes in depth yet (there is a lot :))
I'm afraid I do not fully understand the reasons why you cannot do it like in your 2nd example:
```
Promise.all([
    domtoimage.toPng(element),
    domtoimage.toPng(element),
    ...
]).then(...);
```
That example looks "asynchronous enough" to me actually. One might want to try to identify the hot spots in the code and think about moving them out of the main loop to the web worker, but that is not exactly the thing you're suggesting (please correct me if I'm wrong)

Could you please elaborate on why do we need the `scan()` function and all the complications that arise from that?