@tclausing 2017-06-12
@tsayen Hello. The issue is that I have a batch of DOMs to render, and only one is loaded in the page at a time, so that prevents me from doing as you suggest above. I have to render one, trigger the next one to load, render that one, etc. So the scan function allows me to move on to the next DOM faster - it does only the "clone" parts of the pipeline, and then returns that intermediate data. I can now load the next DOM while the first one is still rendering (by passing the output data from scan() to toPng() or any other existing to* function). That's why I called this approach more helpful in real-world situations, outside of a simple benchmark. Note that even without taking advantage of scan(), all rendering functions get a speedup just from using a native object structure to store DOM info and serialize to xhtml, vs. cloning each element. Maybe I should have submitted two separate PR's...let me know if you like the native object approach but don't like scan(). And thanks for the fantasic library!