@tclausing 2018-08-22
@tamtakoe I don't recall, it's been a while. I'll do that in the next couple days... Though looks like I also need to resolve some conflicts from latest upstream to make sure I am truly isolating the issue. I've also been thinking I should split this PR into two, one to use simple objects to hold the relevant DOM properties instead of actually cloning the DOM; and second to decouple DOM scanning from image generation using web workers. It's possible that just number 1 would fix the problem (unless you already know otherwise?). Just need to find the time...