@pengw00 2021-01-18
hello, is that Promiss.all speedup the process?